Adding the public release of 7.7 onto the GitHub repo.
This version has the following changes and new features:

- A redesigned admin toolbar.
- Lazy-loading of images available to Multiple Image Container plugins.
- Untranslated "placeholder" pages for multi-lingual sites.
- The ability for visitors to see "embed" links.

For more information please see http://zenar.io/zenario-77


Excluding patch releases, this will be the last release for
the Zenario 7 line.

The next major release will be Zenario 8.0; expect to see a change log
for this at some point in the next few weeks.