Various small bugfixes and changes for version 9.2 of Zenario
Quality of life improvements
----------------------------

For newly installed sites, we now set the timeout for user sessions to half an hour (was
15 minutes).
Existing sites can change this by editing their zenario_siteconfig.php file.

If an administrators IP address changes whilst they are logged in, we no longer
automatically log them out.
(Note: This change applies to sites using HTTPS only.)

Where using a jQuery date-picker, in some places we now allow you to manually type the
year you wish to select, instead of having to click the "back" arrow multiple times!

Bugfixes
--------

Fixed a bug where some JavaScript files were not being cached by the browser.

Fixed a bug where some slides in the slideshow created by the installer had corrupted
characters in their titles.

Fix some bad data where some nests/slideshows were not flagged as nests/slideshows in the
database, which lead to some misleading and wrong information appearing in the "where
it's used" columns in Organizer.

Fixed a bug with the translation system, where phrases would not be translated if using
the "link to content item in the current language" feature.

Fixed a database error that sometimes occurred when searching a dataset for a specific
value in MySQL 8.

Fixed a bug with Zenario's file integrity scanning, where CSV files were being incorrectly
blocked.
Fixed a bug with Zenario's file integrity scanning, where some MicroSoft Office files were
being incorrectly blocked.
Also, if an image is blocked by Zenario's file integrity scanning because it has the wrong
extension, we now give better feedback.

Fixed a bug when downloading a file from Organizer, where non-alphanumeric characters in
the filename could sometimes be corrupted.

Fixed a bug in the backup system, where JSON column-types were sometimes not correctly
written to the backup file.

Fixed a bug where the "Flag the first-uploaded image as the featured image" setting for
content items sometimes failed to flag the first-uploaded image.

Fixed a bug when publishing a publicly visible content item, where images on the content
item were not automatically added to the public/images/ directory as intended if they
their privacy was still set to "auto".

Fixed a bug that could occur when duplicating a layout, where the slots in the duplicated
layout were displayed in the wrong order in Organizer.

Fixed a bug that could occur when duplicating a content item on a large site, where it
could take a very long time to duplicate the content item's menu node.

Fixed a bug when trying to run third party programs on the server, where Zenario sometimes
could not find them if their setting was set to look in the PATH variable.

Fixed a bug where some tooltips in admin mode did not have the correct "admin" styling.

Fixed a race-condition that sometimes caused a JavaScript error when showing a plugin with
a ReCAPTCHA.

Fixed a bug with the abstract list plugin, where a user could type a negative number into
the URL to request a negative page number, which would trigger a database error.