Updating the GitHub repo with the latest bugfixes from the 9.3 branch
Fixes for database errors when installing or doing a site reset
---------------------------------------------------------------

The previous release contained a problem with the site reset system, where it was possible
to get a database error whilst doing a site reset. This update will prevent this issue.

We've also fixed a different database error that occurred when trying to install the
Location Manager module when running on PHP 8.1.

Other fixes
-----------

In the plugin settings of the Banner plugin, it's now no longer possible to set mobile
settings for an image before you have selected the image.

Fixed a bug where the unit selector would sometimes not appear when setting a file size
in the site settings.

Several more PHP warnings and notices when running in PHP 8.1 have been fixed.

We've removed the zenario/libs/bower/.htaccess file, as this seems to be causing bugs
and false alarms in GitHub's security advisory system.
If you did have an old link to a library in the bower/ directory, and were relying on this
.htaccess file to correct the broken link for you, then you will need to manually fix any
of these old links.