Updating the GitHub repo with the latest bugfixes from the 9.3 branch
Added some session fixation protection to admin/user logins
-----------------------------------------------------------

Whenever someone logs into Zenario, either as an extranet user or as an admin, we are now
regenerating their session ID.

This has no effect or downside to the user, and helps protect against a session fixation
attack if an attacker is using some other exploit to trigger one.

Other fixes
-----------

PHP 8 has now started issuing a fatal error if you attempt to upload a GIFs image with the
background layer set to an invalid colour. This has caused a lot of confusion, so we've
written a work-around to automatically fix the issue.

Fixed a bug with grid layouts in nests, where if you created a layout without a grouping,
the grid layout inside a nest plugin would fail to work.