Updating the GitHub repo with the latest changes from the 9.3 branch
This update fixes a some issues with the database error reporting tool, and document
downloads.

Automated reporting of database errors
--------------------------------------

The latest versions of PHP come with its own database error reporting tool. However this
has the unfortunately side-effect of stopping Zenario's own database error reporting tool
from working.

We find our own tool better fits our workflow as we rely on the automated emails it sends
to alert us to problems. So we've decided to take the step of turning PHP's version off
for Zenario's scripts.

If this causes an issue for you please leave some feedback on the forums, we could always
consider making this choice a site-config option or a site setting.

Fixes for document downloads
----------------------------

This patch release sees a few fixes for document downloads of documents stored in
Amazon S3.

We've also fixed a bug where document download links would not work if the "Show menu
structure in friendly URLs" option was enabled in the site settings.

Other fixes
-----------

Fixed a bug where the slot controls were missing in some places in admin mode, due to the
security fixes in the previous patch being a little too zealous.

Fixed a mistake where we had linked to the wrong version of the wow.js library in our
package.json file.

Updates for PHP 8.2
-------------------

No more fixes for running on PHP 8.2 coming in the 9.3 branch of Zenario. Support for
running on PHP 8.2 is coming in Zenario version 9.4, to be released shortly.