Updating the GitHub repo with the latest changes from the 9.5 branch
This is purely a bugfix release, with several miscellaneous small fixes.

Updates for minifying skins
---------------------------

After receiving feedback, we've reworked how the "Minify Skin" button works on the
diagnostics screen.
Instead of having to manually press a button, it's now done when you press continue, and
we no longer offer the choice to delay doing this.

We've also fixed a few bugs and inconsistencies, where the ability to do this would
sometimes not appear when a site was in Production mode.

Fixes in visitor mode mode
--------------------------

 * Fixed a bug with the link from the login page to the registration page, where the link
   always went to the page in the default language. It should now go to the page in the
   visitor's current language.
 * Fixed a bug on extranet plugins with a password entry box, where the password
   requirement messages were always in English, and not able to be translated.
 * Fixed a bug when using implied cookie consent together with page caching, where the
   notice message explaining about cookies could sometimes fail close as intended.
 * Fixed an issue in our code where it was possible to trigger a PHP error by hacking the
   URL to the search page.

Fixes in admin mode
-------------------

 * Fixed a bug where, if a site has a custom logo, the error messages shown in staging
   mode were not correct.
 * Fixed a bug where, under certain situations, the admin UI elements would bleed into
   "Preview" mode.
 * Fixed an issue where the WYSIWYG Editor would remove the "rel" attribute from elements.
 * Fixed an issue where the WYSIWYG Editor would corrupt an image's URL if it was used in
   an inline style rule.