This release contains a security patch related to the usage of Twig code in the Twig Snippet plugin, and in the site-wide <head> and <body>.
Critical security patch
-----------------------

The Twig template engine currently has a vulnerability with how some of its filters are
implemented, where it is possible for a designer or an administrator who is aware of the
vulnerability to execute arbitrary CLI code on the server.

This update disables the ability for designers/administrators to call the affected
filters, as a work-around in place of an actual patch for this from Twig's developers.