Updating the GitHub repo with the latest changes from the 9.5 branch.
This release is a bugfix release; with fixes for the page caching system, some fixes for
quality of life features when for logging in as an admin, and some miscellaneous fixes in
admin mode.

Security related fixes
----------------------

We're removed the "Tree Explorer" tool from Organizer, as this was no longer supported and
the code contained a XSS vulnerability.

Fixes in visitor mode
---------------------

 * Fixed a bug with the page caching system, where if you have a document download plugin
   in a nest, which links to a document that's still a first draft, it would crash in
   visitor mode with a PHP fatal error.

Fixes for logging in as an admin
--------------------------------

 * Fixed a bug where the admin login screen would not correctly redirect an administrator
   back to the content item they were previously on, if the page they were previously on
   was served from the page cache.
 * Fixed a bug where the admin login screen would not correctly redirect an administrator
   back to the content item they were previously on, if the administrator had just clicked
   the "AUTHENTICATE" button in the email with their 2FA code.

Fixes in admin mode
-------------------

 * Fixed a bug where the "Save and Next" button for content items was not working for
   content items that did not already have a draft created.