Updating the GitHub repo with the latest changes from the 9.6 branch.
This release addresses an issue with enabling two-factor-authentication for
administrators, and some other miscellaneous fixes.

Fixes for 2FA
-------------

We've fixed a couple of issues that were stopping anyone with the public release of
Zenario from enabling two-factor-authentication for administrators.

Firstly, the section in the `zenario_custom/site_description.yaml` file that needed to be
edited did not have enough guidance as to what that name of the config option actually
was. This has been addressed.

Secondly, we've addressed an issue where the 2FA codes could not actually be sent, due to
a PHP fatal error caused by a missing file in the build.

Fixes in admin mode
-------------------

 * Fixed a rare database error when opening Organizer due to a mistake in a database
   query.
 * Fixed a bug in the public release of Zenario, where pressing the "Suggest" button in
   the Staging Mode admin box caused a PHP fatal error, due to a missing file in the
   build.
 * Fixed an issue where if an inline image was used in a WYSIWYG editor, and if the
   WYSIWYG editor was a library plugin or in a nest, the image could not be centered.
 * Some fixes to the buttons in the View Slots panel in Organizer, where some buttons were
   not appearing in situations where they should do.