Updating the GitHub repo with the latest changes from the 9.6 branch.
This release addresses a couple of issues with using images, and a few other bugs in admin
mode.

Fixes for using images
----------------------

We've fixed a bug where the HTML sanitiser library that we use would not allow large
retina images to be placed in WYSIWYG Editors.
(Their documentation actually recommends keeping a limit, however bumped it up to support
4k images.)

We've also fixed a problem where uploading an image with a % in the name caused various
issues.

Other fixes in admin mode
-------------------------

 * When editing a staging code, if you press the suggest button but then close the Admin
   Box, you'll now see a prompt warning you that it isn't going to be changed if you close
   the box without saving.
 * Fixed a database error when trying to reattach a content item into the menu tree.
 * Fixed a bug when trying to rename a nest or a slideshow, where the system would not let
   you change the case of the characters in the name.
 * Fixed a bug where using the copy/paste buttons in a nest would break the ordinal
   numbers of the plugin buttons in a nest until you rearrange them.