Updating the GitHub repo with the latest changes from the 9.7 branch.
This release addresses a couple of issues with our .htaccess rules, and with running using
MySQL 8.

Fix for a bad .htaccess rule
----------------------------

If you're using the .htaccess rules we provided, you will have included the following rule:

	RewriteRule .*\.svn.*$ - [F,NC]

This contains a mistake in its logic, which will block any file with a timestamp in its
name from the 2nd of May or later.

The rule should be changed to the following instead:

	RewriteRule \.svn/.*$ - [F,NC]

Fixes for timezones in PHP
-----------------

The latest version of PHP removes support for several timezones. Any code that tries to
use these timezones in its logic will trigger a PHP crash.

Our select lists did previously contain some of those removed timezones and were
triggering the crash.

In this patch, we've removed them and tried to replace them with rough equivalents to
preserve functionality.

Fixes for MySQL 8
-----------------

The latest version of MySQL adds some new reserved words. By coincidence, we were using
a few of these as column names, and in a few places we were using them without backticks.
This caused a database error. This update fixes this issue.

Fixes for our caching system
----------------------------

 * Fixed a bug where JavaScript files would sometimes not load on a page when using both the page caching and JavaScript file caching systems at the same time.
 * If a private image is used on a content item, the page/plugin caching system is now completely disabled for that pageload.

Other fixes in admin mode
-------------------------

 * Fixed a PHP warning when migrating a site that uses inline images in WYSIWYG Editors to a later version of Zenario.
 * Fixed a bug where some links to the layouts panel in Organizer were broken.
 * Fixed a bug where you could not remove a plugin in a missing slot, if the slot was on the site-wide header or footer.

Miscellaneous changes
-------------------------

 * We've added two new body CSS classes, body.in_iframe and body.not_in_iframe, that you can use in your CSS code.
 * Any error emails sent by the CMS will now contain the time that the error occurred and the IP address of the person who triggered it.
 * We noticed that the opensource.org website have updated their URLs to a new format. We've updated our links to match.