Copying the public release of 10.2 into the GitHub repo.
Version 10.2 of Zenario is now publicly available, here are the main features.

### Multilingual sites and phrase management

We have improved usability for multi-language sites, by adding a Translations tab to the admin toolbar.

We have incorporated Google Cloud Translation into Zenario, so when working with phrases it is easy to translate a phrase, while still being able to enter translated text manually.

We have split up the Phrases panel in Organizer to be in two panels: for Standard phrase, and for Code-based phrases.

The phrases system now has an "archive" feature. This is useful to remove phrases that you think are no longer needed but would prefer to not completely delete them, just in case they may be needed again.

Importing phrases from Excel sheet should work, and be tolerant of extra (empty) columns

When importing phrases, Zenario now makes sure the module class name is the original module.

We have made it easier to filter in Organizer's content items panels, on Published items.

### Publishing

When publishing a content item, there is now better visual information.

It is possible to choose between immediate publishing, scheduled publishing, and publishing as an unlisted item. There is also information about the item's featured image, its privacy/permissions, language, and categorisation.

We've fixed an issue which could allow a content item that is scheduled for publishing (and therefore locked by that administrator), which could allow it to be unlocked before publishing.

### Images and favicons

We have improved the handling of images in the "special images" area of the library. So it is easy to find and specify key images, such as favicons and social media icons.

It is now possible to specify both a PNG favicon (which is favoured by Google) and an ICO icon (which is favoured by Bing).

In the Library, the Images panels now have a "Not used" filter option. So it is easier to find and tidy up images that are no longer needed.

There is now no concept of an image being archived, and any previously-archived images will now appear again in the image library.

### Menu nodes

We have improved the display of how menu nodes will appear.

We have improved the means of viewing the breadcrumb trail preview.

Fixed a bug whereby a menu node (typically a "Home" of a sub-site) that's used in the Breadcrumbs plugin settings can be deleted, resulting in a broken cross reference. Deleting a referenced menu node is now not allowed.

Fixed a bug whereby if an administrator is on the Diagnostics page and clicks a link that opens a new browser tab, and then opens an Organizer link on that second tab, a warning would occur on the first tab.

There is now a browser reload when an admin attaches a content item to the menu.

On large sites, it was taking a long time to perform a "move" of a menu node. We have greatly improved the efficiency of moving nodes in menu trees of 5,000 nodes and more.

We have updated the interface for the target of a menu node on multi-lingual sites.

### Plugins and nests

Triangle icons for link status now no longer show links in nests and slideshows to the same page.

When editing settings of a plugin, nest or slideshow, there is now more information about where it is used.

### Documents

The documents FEA system is now simplified, as it can now only handle public documents, and not private ones.

When uploading multiple hierarchical documents, they are now sorted alphabetically by default, resulting in better folder organisation. (They can still be dragged into a custom order).

When visitors do a search for document content items, there is now a greatly improved quality textual (full-text) search.

### Extranet features

The Login plugin now has better handling for links after a user has logged in.

### Locations, users and roles

Locations FEA (front-end admin), when viewing a location, more info can now be displayed.

We have improved the edit-user floating admin box so that it's now easier to adding user roles and locations.

#### User timers

When cancelling a user's timer, and the timer has a grace period, the actions for the end of the grace period are now processed.

### Forms

We have improved the Forms editor, to better organise the decorative and custom form field types which an admin can add. Also, it is now possible that when adding a subheading, it's possible to choose what tag (p/h1/h2/h3/h4/h5/h6) is used.

### Organizer

When "fixing" an error recorded in the error log, Zenario now deals correctly with aliases that had an argument passed to to them after a "?" in the URL.

In the 404 Error log panel, there is now a warning if the error log is bigger than 10,000 entries, and there are now more options for how long to store it.

There is now a "Security" site settings panel, which allows more control over security warnings when an administrator logs in.

There are better controls on the Scheduled Tasks panel, and certain actions to enable or run a task can now only be done while scheduled tasks are set up correctly and running.

When adding something to a slot on a content item using the slot controls, there is now a simpler interface to let the admin choose a plugin, a nest or a slideshow.

We have fixed most warnings from Google Maps.

There is now a control for determining the day of the week to show first when displaying a calendar.

*These release notes for Zenario 10.2 can be found at [zenar.io/zenario-102](https://zenar.io/zenario-102).*