Copying the public release of 10.3 into the GitHub repo.
Version 10.3 of Zenario is now publicly available, here are the main features.

Creating and editing content items
----------------------------------

### Creating content items

We have simplified the content item admin box for creating and editing content
items. For content such as HTML page content items, the box no longer contains
the main content tab; but for items such as News and Blog, the tab remains, so
that these things can be created quickly. These settings can all be customised
in Organizer->Content type settings.

### Duplicating and translating

When duplicating a content item, or translating it into a different language,
there is now a 2-column box, making the work far more intuitive.

This extends to the "Slots" tab of the duplication box, which is now easier to
use.

We have added a "Translate with Google" feature when duplicating a content item
into a new language.

Menu navigation
---------------

When generating the menu on large sites, there are now considerable speed
improvements in the drawing algorithm, resulting in quicker page generation.

We also have greatly speeded up the menu for mobile with push-pull, now
optimised for large sites.

In the menu hierarchy view in Organizer, it's now possible to select multiple
menu nodes at once, and make them visible or invisible. (Invisible menu nodes
have a breadcrumb trail when seen by a visitor, but don't appear in regular
menus).

The "Main" menu section can now no longer be renamed.

Search and lists
----------------

In the Advanced Search module, there are now more controls for adjusting
weightings.

There is also a new searchable cache database table, and this is used by the
module to provide much faster search results.

Content Summary List plugins can now sort in reverse alphabetical order on
browser title.

General management
------------------

When in Organizer, and there is an empty panel, there is now more friendly
information about what that panel may contain, and how to create it. Once it
contains some records, the same information is available via the Help button.

Content items now have a "last activity" time (as opposed to time of version
creation, last version edit, etc.), so it's easier to see what has been worked
on most recently.

When picking content items, HTML pages are now always listed first.

We have improved the file handling for document content items, so that they now
have their files stored in the public directory with a fixed 5-letter
subdirectory name, thus helping search engine optimisation. (Applies to public
documents).

Content item browser titles can now be up to 250 characters. This is determined
in the content type settings.

We have rearranged the administrator permissions interface, and made it easier
to create a regular client site administrator with a standard set of
permissions.

We have removed support for setting background images in content items. This
should be done using CSS instead, and its removal saves on clutter within admin
boxes.

We have simplified the Modules panel, so the first Organizer panel of modules
now just shows those modules that are running and suspended.

The XML sitemap now has styling, which does not change functionality but looks
more professional.

Aliases/spare aliases (redirects)
---------------------------------

Zenario now permits requests for `.doc`, `.docx`, `.pdf` and `.zip` files to be
redirected to content items via a spare alias.

Forms
-----

Version 10.3 introduces a much-requested feature where the destination email
address(es) of a form can depend on a select list in the form. For example, when
a user chooses the nature of their enquiry, the form response data can be
emailed automatically to the correct department.

There is now a facility so that when a visitor is entering an email, there is
confirmation by asking the user to enter their email address twice.

There is now a warning when a form is used in a regular (non-Ajax) nest, as they
do not work well in nests.

Email and newsletters
---------------------

When creating a newsletter, there is now a more friendly wizard-like interface.
It allows the administrator to pick a template, a previously sent newsletter, or
to start from scratch with an empty one.

Any images used in emails and newsletters are now sent in their original PNG or
JPEG format, and not in the modern WebP format. WebP had been implemented to
improve performance, but it was causing compatibility issues on older email
clients and webmail clients. Emails with images should now display correctly.

Job vacancies content
---------------------

When posting job vacancies (using the Job Vacancy content type), and using a
response form, you can now have a specific response email address for that
vacancy, to override the one determined on the form.

Social media
------------

We've improved HTML tag handling in the page source, with better control over
social media meta tags. The keywords tag is now only included when there are
some keywords in a page, and there's more control over the generator tag.

Multi-lingual phrase management
-------------------------------

When viewing visitor phrases in Organizer, there is now better information about
when and where the phrases was seen and therefore created.

Password-protected (extranet) site management
---------------------------------------------

There is now support for weak user passwords, which may be useful in an intranet
type environment, such as where Zenario is run in a firewalled environment.

There is also now support for setting a password strength for administrators
that is independent of the password strength for users.

For User Timers, there is now a safeguard in Timers FEA module which prevents a
timer being granted where it relates to a group of which the user is already a
member.

We now show better information on a content item about which group(s) it is
shared to, by showing a small coloured disc on the admin toolbar with the number
of groups.

The edit-permissions admin box now shows a summary of how many people are
extranet users, in each group, and smart group.

When converting a Contact to a User, there is now a more accurate email template
used for the email, and it now contains the user's newly-generated password.

We have improved Organizer searching in the users and contacts panel.

Tools for system administrators
-------------------------------

There is now a "moratorium" mode for a Zenario site, so that access for all
local administrators can be temporarily blocked. This is helpful when upgrading
a site, and the system administrator needs to make sure content does not get
changed during the upgrade process.

The hierarchical documents system now has a check during the admin
diagnostics/login page, with a "Repair" button to repair problems in the public
folder.

We have removed support for Dropbox integration as this was not well used.

MIC (Multiple Image Container plugin) images are now stored in their own folder
inside public directory

Tools for developers
--------------------

Version 10.3 sees some more useful tools added to our debug menus.

For most things that are powered by TUIX, you can now get a list of the source
files that were used, and copy a prompt to ask an AI assistant about what you
are looking at.

We are now also displaying a few useful variables about the state of the
application, to help with debugging.

These release notes for Zenario 10.3 can be found at https://zenar.io/zenario-103